# perf: Speed improvements for moduleview get