--- a/frappe/auth.py
+++ b/frappe/auth.py
@@ -142,6 +142,7 @@ class LoginManager:
 		self.validate_hour()
 		self.get_user_info()
 		self.make_session()
+		self.setup_boot_cache()
 		self.set_user_info()
 
 	def get_user_info(self, resume=False):
@@ -149,6 +150,11 @@ class LoginManager:
 			["user_type", "first_name", "last_name", "user_image"], as_dict=1)
 
 		self.user_type = self.info.user_type
+
+	def setup_boot_cache(self):
+		frappe.cache_manager.build_table_count_cache()
+		frappe.cache_manager.build_domain_restriced_doctype_cache()
+		frappe.cache_manager.build_domain_restriced_page_cache()
 
 	def set_user_info(self, resume=False):
 		# set sid again