--- a/frappe/utils/user.py
+++ b/frappe/utils/user.py
@@ -157,10 +157,11 @@ class UserPermissions:
 				self.can_read.remove(dt)
 
 		if "System Manager" in self.get_roles():
-			docs = [x["name"] for x in frappe.get_all("DocType", "name")]
-			for docname in docs:
-				if frappe.get_meta(docname, cached=True).allow_import == 1:
-					self.can_import.append(docname)
+			docs = frappe.get_all("DocType", {'allow_import': 1})
+			self.can_import += [doc.name for doc in docs]
+
+			customizations = frappe.get_all("Property Setter", fields=['doc_type'], filters={'property': 'allow_import', 'value': 1})
+			self.can_import += [custom.doc_type for custom in customizations]
 
 		frappe.cache().hset("can_import", frappe.session.user, self.can_import)