The PR replaces small queries in loops and replaces them with either cached data or consolidated queries.

The following are the changes

1. Caching API for Table Counts, DocTypes and Pages restricted to domain in `cache_manager`
1. Build the above mentioned cache on login in `auth.py`
1. Added rebuilding of the cache in hooks.py on insert of any document (This cache is utilized in `moduleview.py` for setting doctype count and for filtering the doctypes based on domain)
1. Replaced the `get_meta` in `build_permissions` of `user.py` with two specific queries... The computational load of getting these is quite higher than the query time of fetching them from DB

Improvements for **Administrator** and **System Manager**

| Query                | Baseline | Optimized | Improvement |
|----------------------|----------|-----------|------------|
| Fresh Boot                 | 8900ms     | 560ms       | ~93%       |
| Accounts Module Page | 2012ms     | 392ms       | ~80%       |
| Website Module Page  | 1362ms     | 368ms       | ~70%       |


Improvements for **Users without System Manager Role**

| Query                | Baseline | Optimized | Improvement |
|----------------------|----------|-----------|------------|
| Fresh Boot                 | ~900ms     | ~900ms       | No Improvments       |
| Accounts Module Page | 869ms | 240ms | ~ 72%    |
| Website Module Page  | 292ms     | 210ms       | ~28%      |